Type cast intermediate bit shift in pixel buffer function
- Again another runtime error according to ASAN
- Did a bit of research and both the operands in a bitshift operation
  follow C's integer promotion rules with the result having the same
  type as the left operand, in this case uint8_t is promoted to an
  int (and shifting into a signed bit is undefined behaviour)
- Minor change which type casts the bit shift specifically

+++ src/ppu.c
@@ -611,7 +611,7 @@ void set_rgba_pixel_in_buffer(uint32_t* pixel_buffer, unsigned max_width
                              , unsigned int x_pos, unsigned int y_pos
                              , unsigned int rgb, uint8_t alpha)
 {
-	pixel_buffer[x_pos + (max_width * y_pos)] = (alpha << 24) | rgb;
+	pixel_buffer[x_pos + (max_width * y_pos)] = ((uint32_t) alpha << 24) | rgb;
 }
